requirements_airflow.txt
apache-airflow==2.8.1
apache-airflow-providers-sqlite==3.6.0
